Overhaul management of org-agenda files
Org-agenda files are now separated from org-mode files found in
`org-directory', which are stored in `org-files'. Org-agenda files are
derived from `org-files' depending on the presence of a certain
filetag defined in org-agenda-file-tag.

While org-agenda is now built with a potentially smaller subset of
`org-files' for reasons of speed, other applications such as helm-rifle
may process all org-mode files that are found in `org-directory'.